requirements.txt
aiogram==2.11.2
aiohttp>=3.7.3
aioredis==1.3.1
python-dotenv
requests
beautifulsoup4
lxml